Refactor document types to separate types file and update imports
